Changes by 이종원

+++ src/main/java/com/sparta/bookflex/domain/basket/dto/BasketResDto.java
@@ -3,17 +3,19 @@ package com.sparta.bookflex.domain.basket.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class BasketResDto {
 
     private Long basketId;
     //private PhotoImage photoImage;
     private String bookName;
-    private int price;
+    private BigDecimal price;
     private int quantity;
 
     @Builder
-    public BasketResDto(Long baseketId, String bookName, int price, int quantity) {
+    public BasketResDto(Long baseketId, String bookName, BigDecimal price, int quantity) {
         this.basketId = baseketId;
         //this.photoImage = photoImage;
         this.bookName = bookName;
